updated to play 2.3 removed local javascript files, those are now pulled from web jars
