+++ app/views/index.scala.html
@@ -1,10 +1,9 @@
 @()
 
-@main("Dashboard") {
+@main("javascripts/main.js") {
   <!-- Repeat by Dimitry Sokolov from The Noun Project -->
   <!-- Moon by Jory Raphael from The Noun Project -->
   <!-- Smart Phone by Thomas Le Bas from The Noun Project -->
   <!-- Input by Icons8 from The Noun Project -->
-  <script src="@routes.Assets.at("javascripts/dashboard.js")" type="text/jsx"></script>
   <div id="content"></div>
 }
