+++ app/views/main.scala.html
@@ -1,16 +1,13 @@
-@(title: String)(content: Html)
+@(jsFile: String)(content: Html)
 
 <!DOCTYPE html>
 
 <html>
     <head>
-        <title>@title</title>
-        <link rel="stylesheet" media="screen" href="@routes.Assets.at("stylesheets/dashboard.css")">
-        <link rel="shortcut icon" type="image/png" href="@routes.Assets.at("images/favicon.png")">
-        <script src="@routes.Assets.at("javascripts/jquery-1.9.0.min.js")" type="text/javascript"></script>
-        <script src="@routes.Assets.at("javascripts/moment.min.js")" type="text/javascript"></script>
-        <script src="http://fb.me/react-with-addons-0.9.0.js"></script>
-        <script src="http://fb.me/JSXTransformer-0.9.0.js"></script>
+        <title>Dashboard</title>
+        <link rel="stylesheet" media="screen" href="@routes.Assets.versioned("stylesheets/dashboard.css")">
+        <link rel="shortcut icon" type="image/png" href="@routes.Assets.versioned("images/favicon.png")">
+        <script data-main="@routes.Assets.versioned(jsFile)" type="text/javascript" src="@routes.Assets.versioned("lib/requirejs/require.js")"></script>
     </head>
     <body>
         @content
