+++ conf/routes
@@ -14,3 +14,4 @@ POST  /setDevice/    controllers.Application.setDevice()
 
 # Map static resources from the /public folder to the /assets URL path
 GET   /assets/*file     controllers.Assets.at(path="/public", file)
+GET   /assets/*file                            controllers.Assets.versioned(path="/public", file: Asset)
